change double to float for featue vectors and fix errors

+++ Assets/Scripts/QuestUtils/BasicQuest.cs
@@ -1,4 +1,7 @@
+using Microsoft.Geospatial;
+using UnityEngine;
+
 public class BasicQuest : Quest {
-    public BasicQuest(string label, Texture2D referenceImage, double[] featureVector) 
+    public BasicQuest(string label, Texture2D referenceImage, float[] featureVector) 
         : base(label, referenceImage, featureVector) {}
 }
